% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\name{rmacrolite-system-variables}
\alias{rmacrolite-system-variables}
\title{Windows System Environment Variables used by rmacrolite}
\description{
It is possible, but not compulsory, to define 
 Windows System Environment Variables, 
 \code{rmacrolite_macro_path}, \code{rmacrolite_macro_exe}, 
 and/ or \code{rmacrolite_macro_exeparfile}, with the default 
 value to be used by the package \code{rmacrolite} for 
 the path to the folder where MACRO or MACRO In FOCUS is 
 installed, the name of MACRO or MACRO In FOCUS executable 
 and the name of the so-called exeparfile executable, 
 respectively. The path to MACRO executable is the most 
 important, as it can vary between computers.
}
\details{
When these Windows System Environment Variables have been 
 set, it is not necessary to use the function 
 \code{\link[rmacrolite:rmacroliteSetModelVar-methods]{rmacroliteSetModelVar}} in your code, 
 as the package will automatically find the relevant values 
 that need to be used.

These variables may be especially useful when trying to 
 build the package from source and run \code{R CMD check}.

If you want to check if these System variable have been 
 set and what is their current value, you can type:
 \code{Sys.getenv("rmacrolite_macro_path")} or 
 \code{Sys.getenv("rmacrolite_macro_exe")} or 
 \code{Sys.getenv("rmacrolite_macro_exeparfile")}.
 A value of \code{""} indicates that the system variables 
 do not exist.

You (most likely) need administrator rights on your 
 computer to set these variables. Or you need your System 
 admin to set them. In Windows 10, the variables can be 
 set as: Windows Start menu > Settings > System > About > 
 Related settings - system info > Advanced System Settings 
 > Advanced tab > "Environment variables..." > System variables.
 Alternatively type 
 \code{shell.exec("SystemPropertiesAdvanced.exe")} in 
 R command prompt, to reach the Advanced System Settings.
 If these variables don't exist, they need to be created.
}
\examples{
Sys.getenv("rmacrolite_macro_path", unset = NA_character_ )
Sys.getenv("rmacrolite_macro_exe", unset = NA_character_ ) 
Sys.getenv("rmacrolite_macro_exeparfile", unset = NA_character_ ) 

#   If NA is returned, the System Variable has not been set
}
\seealso{
\code{\link[rmacrolite:rmacroliteSetModelVar-methods]{rmacroliteSetModelVar}} and 
 \code{\link[rmacrolite]{rmacroliteGetModelVar}}.
}
