% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\name{rmacroliteChange1Param}
\alias{rmacroliteChange1Param}
\title{Change the value of one parameter from an imported MACRO parameter file (PAR)}
\usage{
rmacroliteChange1Param(x, pTag, type = NULL, value,
  tagNb = NA_integer_)
}
\arguments{
\item{x}{A \code{macroParFile} object, as obtained with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}.}

\item{pTag}{Single character string. Text string containing the parameter 
value that should be searched and replaced in the PAR file, 
with the parameter value replaced by the string \code{\%s}.
NOTE: CASE SENSITIVE!}

\item{type}{Single character string. Parameter category (category's header 
in the PAR file).}

\item{tagNb}{Vector of integer values. If several rows have the same 
\code{pTag}, indicates which one to change (1 for the 1st, 2
for the 2nd, etc.).}

\item{values}{New values for the parameter in \code{pTag}. Single or vector 
of integer or real or character (etc.) value(s).}
}
\value{
Returns vector of \bold{character} strings.
}
\description{
Change the value of one parameter from an imported MACRO 
 parameter file (PAR). If several rows are found that match 
 the parameter tag, all values are changed.
}
\examples{

library( "rmacrolite" )

#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" )  

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 

#   Current value of RUNID
as.numeric( rmacroliteGet1Param( 
     x    = par_file, 
     pTag = "RUNID\\t\%s", 
     type = "HEAD" ) )

#   Change the RUNID
par_file <- rmacroliteChange1Param( 
    x     = par_file, 
    pTag  = "RUNID\\t\%s", 
    type  = "HEAD", 
    value = 2L ) 

#   Check new value
new_run_id <- as.numeric( rmacroliteGet1Param( 
     x    = par_file, 
     pTag = "RUNID\\t\%s", 
     type = "HEAD" ) ) 

new_run_id

#   Internal checks
if( !(new_run_id == 2L) ){
    stop( "Test of rmacroliteChange1Param() failed" )
}   

#   Clean-up
rm( par_file_path, par_file, new_run_id )
}
\keyword{internal}
