% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteClimateFiles}
\alias{rmacroliteClimateFiles}
\alias{rmacroliteClimateFiles.macroParFile}
\title{Fetch the name and path of the rainfall and weather data bin-files in an imported MACRO par-file}
\usage{
rmacroliteClimateFiles(x, ...)

\method{rmacroliteClimateFiles}{macroParFile}(x, check = TRUE, ...)
}
\arguments{
\item{x}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{check}{Single logical values. If \code{TRUE} (the default), 
the function checks that the two files 
exists and stops when they don't.}
}
\value{
A vector with two named character strings. \code{rain}, 
 the name and path to the rainfall file, and \code{met}, 
 the name and path to the weather file.
}
\description{
etch the name and path of the rainfall and weather data 
 bin-files in an imported MACRO par-file
}
