% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteCropUptF}
\alias{rmacroliteCropUptF}
\alias{rmacroliteCropUptF.macroParFile}
\alias{rmacroliteCropUptF<-}
\alias{rmacroliteCropUptF<-.macroParFile}
\title{Fetch or set the crop uptake factor in an imported MACRO par-file}
\usage{
rmacroliteCropUptF(x, ...)

\method{rmacroliteCropUptF}{macroParFile}(x, ...)

rmacroliteCropUptF( x, ... ) <- value

\method{rmacroliteCropUptF}{macroParFile}(x, ...) <- value
}
\arguments{
\item{x}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{value}{Single numeric-value. New value of the crop uptake factor 
to be set in the imported par-file (\code{x}).}
}
\value{
Single numeric-value. Current value of the crop uptake 
 factor in the imported par-file (\code{x}).
}
\description{
Fetch or set the crop uptake factor in an imported MACRO 
 par-file. The crop uptake factor is FSTAR in MACRO 
 par-files.
}
\examples{

library( "rmacrolite" ) 

# # Setup MACRO directory (if needed)
# rmacroliteSetModelVar( "C:/swash/macro" )



#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" ) 

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 



#   Fetch the current parametrization
rmacroliteCropUptF( x = par_file ) 
    # [1] 0.9

#   Modify the parameter
par_file2 <- par_file
rmacroliteCropUptF( x = par_file2 ) <- 0.5 

#   Check the result
cuf <- rmacroliteCropUptF( x = par_file2 )
cuf 
    # [1] 0.5

#   Internal control
if(cuf != 0.5 ){ 
    stop( "Test of rmacroliteCropUptF() failed" ) } 

#   Clean-up
rm( par_file_path, par_file, par_file2, cuf  )

}
