% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteDiffCoef}
\alias{rmacroliteDiffCoef}
\alias{rmacroliteDiffCoef.macroParFile}
\alias{rmacroliteDiffCoef<-}
\alias{rmacroliteDiffCoef<-.macroParFile}
\title{Fetch or set the substance diffusion coefficient [m2/s] in an imported MACRO par-file}
\usage{
rmacroliteDiffCoef(x, ...)

\method{rmacroliteDiffCoef}{macroParFile}(x, ...)

rmacroliteDiffCoef( x, ... ) <- value

\method{rmacroliteDiffCoef}{macroParFile}(x, ...) <- value
}
\arguments{
\item{x}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{value}{Single numeric-value. New value of the substance diffusion 
coefficient [m2/s] to be set in the imported par-file 
(\code{x}).}
}
\value{
Single numeric-value. Current value of the substance 
 diffusion coefficient [m2/s] in the imported par-file 
 (\code{x}).
}
\description{
Fetch or set the substance diffusion coefficient [m2/s] 
 in an imported MACRO par-file. The diffusion coefficient 
 is DIFF in MACRO par-files.
}
\examples{

library( "rmacrolite" ) 

# # Setup MACRO directory (if needed)
# rmacroliteSetModelVar( "C:/swash/macro" )



#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" ) 

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 



#   Fetch the current parametrization
rmacroliteDiffCoef( x = par_file ) 
    # [1] 5.2e-10

#   Modify the parameter
par_file2 <- par_file
rmacroliteDiffCoef( x = par_file2 ) <- 5.00E-10

#   Check the result
dc <- rmacroliteDiffCoef( x = par_file2 )
dc 
    # [1] 5e-10

#   Internal control
if(dc != 5.00E-10 ){ 
    stop( "Test of rmacroliteDiffCoef() failed" ) } 

#   Clean-up
rm( par_file_path, par_file, par_file2, dc  )

}
