% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteExport}
\alias{rmacroliteExport}
\alias{rmacroliteExport.macroParFile}
\alias{rmacroliteExport.macroSimResults}
\alias{rmacroliteExport.macroSimResultsList}
\title{Export MACRO parameter sets or MACRO simulation results}
\usage{
rmacroliteExport(x, where = NULL, ...)

\method{rmacroliteExport}{macroParFile}(x, where = NULL, ...)

\method{rmacroliteExport}{macroSimResults}(x, where = NULL,
  overwrite = FALSE, ...)

\method{rmacroliteExport}{macroSimResultsList}(x, where = NULL,
  overwrite = FALSE, ...)
}
\arguments{
\item{x}{A \code{macroSimResults}, \code{macroSimResultsList} or 
\code{macroParFile}-object, 
containing one (or several) simulation parameter set(s) or 
result(s). \code{macroParFileList} are not supported yet.}

\item{where}{Single character string. If \code{x} is a 
\code{macroParFile}-object, absolute or relative path 
of the folder in which the simulation-parameter(s) should be 
written. If \code{where} is \code{NULL}, a temporary working 
directory will be created where the simulations will be written. 
If \code{x} is a \code{macroSimResults} \code{macroSimResultsList}, 
absolute or relative path AND file name prefix (without 
extension) of the file in which the simulations shall be 
exported.}

\item{\dots}{Additional parameters passed to \code{\link[utils:write.table]{write.csv}} 
or \code{\link[rmacrolite:rmacroliteExportParFile-methods]{rmacroliteExportParFile}}.}

\item{overwrite}{Single logical value. If \code{TRUE}, will overwrite the files 
if they already exist.}
}
\value{
TO DO: COMPLETE THIS.
}
\description{
Export MACRO parameter sets or MACRO simulation results
}
\seealso{
\code{\link[rmacrolite:rmacroliteExportParFile-methods]{rmacroliteExportParFile}}.
}
