% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\name{rmacroliteGetModelVar}
\alias{rmacroliteGetModelVar}
\title{Set the absolute path of the folder in which MACRO (or MACRO In FOCUS) executable is installed.}
\usage{
rmacroliteGetModelVar()
}
\value{
Invisibly returns a \code{\link[base]{list}} with 3 named 
 items: \code{path}, \code{exe} and \code{exeparfile}, 
 set to the values found out by the function.
}
\description{
Set the absolute path of the folder in which MACRO (or 
 MACRO In FOCUS) executable is installed. as well as 
 the name of MACRO executable and the name of the exeparfile 
 executable.
 
 Regarding the path to MACRO-folder, the function 
 proceeds as follow: If a value is given (argument 
 code{path}, see below), it is used to set MACRO-path in 
 the package-option \code{macro_path} (see 
 \code{\link[rmacrolite]{rmlPar}}), value that is then used 
 by other function in this package. If no value is given 
 (argument code{path} left to \code{NULL}), the function 
 will first search for a Windows Environment Variable 
 (System Variable) called \code{rmacrolite_macro_path}, 
 and use this value if it exists, to set \code{macro_path} 
 in \code{\link[rmacrolite]{rmlPar}}. See 
 \code{\link[rmacrolite]{rmacrolite-system-variables}}. 
 Finally, if neither the argument 
 code{path} is set, nor a System Variable called 
 \code{rmacrolite_macro_path}, the function will try the 
 factory-default path to MACRO In FOCUS (type 
 \code{getRmlPar("macro_path_default")} to find out).
 
 The principle is the same for MACRO executable (argument 
 \code{exe}, System Variable \code{rmacrolite_macro_exe}, factory 
 default set in \code{getRmlPar("macro_exe_default")}) 
 as well as for the exeparfile executable (argument 
 \code{exeparfile}, System Variable 
 \code{rmacrolite_macro_exeparfile}, factory default set in 
 \code{getRmlPar("macro_exeparfile_default")}).
}
\examples{

library( "rmacrolite" )

rmacroliteGetModelVar()

    # $`path`
    # [1] "C:\\\\swash\\\\macro"

    # $exe
    # [1] "Macro52Model.exe"

    # $exeparfile
    # [1] "exeparfile.exe"
}
\seealso{
\code{\link[rmacrolite:rmacroliteSetModelVar-methods]{rmacroliteSetModelVar}} and 
 \code{\link[rmacrolite]{rmacrolite-system-variables}}.
}
