% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteImportParFile}
\alias{rmacroliteImportParFile}
\alias{rmacroliteImportParFile.default}
\title{Imports parameters from one or several MACRO parameter-file(s)}
\usage{
rmacroliteImportParFile(file, climate = TRUE, verbose = 1L, ...)

\method{rmacroliteImportParFile}{default}(file, climate = TRUE,
  verbose = 1L, ...)
}
\arguments{
\item{file}{Single character string or vector of character strings.
Name(s) of the MACRO In FOCUS parameter file(s) (par-file) 
to be imported. The file(s) should be located in the same 
folder as the MACRO In FOCUS executable 
(see \code{\link[rmacrolite:rmacroliteSetModelVar-methods]{rmacroliteSetModelVar}}), or in a subfolder
in this folder (in this case indicate the relative path, 
not the full path).

Notice that R file separator is a slash (\code{/}), or a double 
slash, but not a single backslash (although double backslash 
would work).}

\item{climate}{Single logical value. If \code{TRUE} (the default), the 
function checks that the climate files exists and stop 
if they don't.}

\item{verbose}{Single integer value. If set to a value \code{< 1}, 
the program is silent (except on errors or warnings). If 
set to \code{1}, the program outputs messages. Values 
\code{> 1} may also activate messages from lower level 
functions (for debugging purpose).}

\item{\dots}{Additional parameters passed to \code{\link[base]{readLines}}.}
}
\value{
Returns a 'macroParFile' object
}
\description{
Imports parameters from one or several MACRO parameter-file(s)
}
\examples{

library( "rmacrolite" )

#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" )  

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 

#   Clean-up
rm( par_file_path, par_file )

}
