% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteLayers}
\alias{rmacroliteLayers}
\alias{rmacroliteLayers.macroParFile}
\title{Fetch information (depths, thicknesses) on the layers in an imported MACRO par-file}
\usage{
rmacroliteLayers(x, ...)

\method{rmacroliteLayers}{macroParFile}(x, ...)
}
\arguments{
\item{x}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}
}
\value{
Returns a list with two items, \code{"layer"} 
 (layer-specific parameters) and \code{"site"} 
 (site-specific parameters, i.e. parameters that do 
 not vary with depth or with the crop). \code{"layer"} is 
 a \code{\link[base]{data.frame}} with the following 
 columns \code{layer_no}, \code{thick_cm}, 
 \code{depth_from_cm}, \code{depth_to_cm} and  
 \code{nb_num_layers} (the thickness [cm], the upper and 
 lower depths [cm] and the number of numerical layers, 
 respectively). \code{"site"} is a vector of named 
 numeric values, with the following itemps 
 \code{nb_horizons1} (the number of layers/ horizons; 
 parameter NHORIZON, header of the par-file), 
 \code{nb_horizons2} 
 (parameter NHORIZONS, PHYSICAL PARAMETERS section of the par-file) 
 \code{nb_horizons3} 
 (parameter NHORIZONS, PROPERTIES section of the par-file) 
 and \code{nb_num_layers} (the total number number of 
 numerical layers; parameter NLAYER in the par-file 
 header).
}
\description{
Fetch information on the layers (i.e "horizons") in an 
 imported MACRO par-file: thicknesses [cm], start and end 
 dephs [cm], number of numerical layers. These values are 
 based on, or corresponds to the parameters NHORIZON 
 (par-file header), NHORIZON (PROPERTIES section), NLAYER 
 (par-file header), HTICK and NLAYER (PROPERTIES ROPERTIES 
 section) in MACRO par-files.
}
