% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteMacroVersion}
\alias{rmacroliteMacroVersion}
\alias{rmacroliteMacroVersion.character}
\title{Fetch MACRO version from the folder where it is installed}
\usage{
rmacroliteMacroVersion(path = character(0), ...)

\method{rmacroliteMacroVersion}{character}(path = character(0), ...)
}
\arguments{
\item{path}{A single character-strings. Path to the directory where 
MACRO (or MACRO In FOCUS) is installed and where the 
model version is to be found. When \code{path} equal to 
\code{character(0)} (the default), it is retrieved 
automatically using 
\code{\link[rmacrolite]{rmacroliteGetModelVar}[["path"]]}.}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}
}
\value{
TO BE WRITTEN
}
\description{
Fetch MACRO version from the folder where it is installed
}
\examples{
rmacroliteMacroVersion()
}
