% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteRun}
\alias{rmacroliteRun}
\alias{rmacroliteRun.character}
\alias{rmacroliteRun.macroParFile}
\alias{rmacroliteRun.macroParFileList}
\title{Run one or several MACRO Simulation(s)}
\usage{
rmacroliteRun(x, f = NULL, export = TRUE, verbose = 1L, ...)

\method{rmacroliteRun}{character}(x, f = NULL, export = TRUE,
  verbose = 1L, ...)

\method{rmacroliteRun}{macroParFile}(x, f = NULL, export = TRUE,
  verbose = 1L, indump = TRUE, run = TRUE, rename = FALSE, ...)

\method{rmacroliteRun}{macroParFileList}(x, f = NULL, export = TRUE,
  verbose = 1L, ...)
}
\arguments{
\item{x}{A \code{macroParFile}- or a \code{macroParFileList}-object, 
containing one or several simulations to be ran. Alternatively, 
a single character string giving the name of and optionally 
the path to a par-file to be imported and simulated. When 
\code{x} is a single character string and \code{export} 
is \code{FALSE}, \code{f} should be \code{NULL} or identical 
to \code{x}, for consistency. When \code{x} is a single 
character string and \code{export} is \code{TRUE}, 
\code{f} should be \code{NULL} or set 
to a name different than the one in \code{x}, in order 
not to overwrite the original file.}

\item{f}{Single character string. When \code{export} is \code{TRUE}, 
name (without path) of the par-file where the simulations 
parameters should be written. 
If \code{NULL}, a name will be attributed using 
the template given by 
\code{\link[rmacrolite]{getRmlPar}("fileNameTemplate")} 
and the RUNID contained in the par-file. When \code{export} 
is \code{FALSE}, name (without path) of an existing par-file 
that should be used to run the simulation. In all cases 
the location of the par-file is the directory in which 
MACRO is installed, as given by 
\code{\link[rmacrolite]{rmacroliteGetModelVar}()[["path"]]}.}

\item{export}{Single logical value. If \code{TRUE} (the default), the 
par-file is exported prior to the simulation. If 
\code{FALSE}, it is assumed that the par-file to be simulated 
already exists and its name and optionally location is 
indicated by}

\item{verbose}{Single integer value. If set to a value \code{< 1}, 
the program is silent (except on errors or warnings). If 
set to \code{1}, the program outputs messages. Values 
\code{> 1} may also activate messages from lower level 
functions (for debugging purpose).}

\item{\dots}{Additional parameters passed to specific methods.}

\item{indump}{Single logical value. If \code{TRUE} (the default), 
the so called \code{indump.tmp} parameter file is produced. 
Must be \code{TRUE} when \code{run} is \code{TRUE}.}

\item{run}{Single logical value. If \code{TRUE} (the default), the 
simulation is run. If \code{FALSE}, only the par-file is 
exported (when \code{export} is \code{TRUE}) but the 
simulation is not run.}

\item{rename}{Single logical value. If \code{TRUE}, the bin-file output 
by MACRO is renamed automatically, using the template 
in \code{\link[rmacrolite]{getRmlPar}("fileNameTemplate")}.
If \code{FALSE} (the default), the bin-file output 
by MACRO is not renamed.}
}
\value{
Returns a \code{\link{data.frame}} with the simulation results, 
 a column \code{date} and if relevant a column \code{runId}.
}
\description{
Run one or several MACRO Simulation(s)
}
\examples{

library( "rmacrolite" ) 

# # Setup MACRO directory (if needed)
# rmacroliteSetModelVar( "C:/swash/macro" )



# Import a par-file

#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182_1910-1911.par", 
    package = "rmacrolite" )  

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 



# Run the simulation
out <- rmacroliteRun( x = par_file ) 



# Inspect the results

#   View the water and solute balance as output by MACRO
#   attributes( out ) # full list of attributes
attr( x = out, which = "waterSoluteBalance" )

library( "macroutils2" )

macroPlot( x = out[, c( "Date", "WOUT_99-100_100", 
    "SFLOW_99-100_100" ) ], gui = FALSE, subPlots = TRUE ) 


}
