% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteRunId}
\alias{rmacroliteRunId}
\alias{rmacroliteRunId.macroParFile}
\alias{rmacroliteRunId.macroParFileList}
\alias{rmacroliteRunId<-}
\alias{rmacroliteRunId<-.macroParFile}
\alias{rmacroliteRunId<-.macroParFileList}
\title{Fetch or set the simulation ID (RUNID) of one or more imported MACRO simulation parameter sets}
\usage{
rmacroliteRunId(x, ...)

\method{rmacroliteRunId}{macroParFile}(x, ...)

\method{rmacroliteRunId}{macroParFileList}(x, ...)

rmacroliteRunId( x, ... ) <- value

\method{rmacroliteRunId}{macroParFile}(x, ...) <- value

\method{rmacroliteRunId}{macroParFileList}(x, ...) <- value
}
\arguments{
\item{x}{A \code{macroParFile} object, containing one simulations 
whose simulation ID (RUNID) should be fetched or set.}

\item{\dots}{Additional parameters passed to specific methods.}

\item{value}{A single or vector of integer. New value(s) for the RUNID}
}
\value{
WRITE DESCRIPTION HERE.
}
\description{
Fetch or set the simulation ID (RUNID) of one or more imported 
 MACRO simulation parameter sets
}
\examples{

library( "rmacrolite" ) 

# # Setup MACRO directory (if needed)
# rmacroliteSetModelVar( "C:/swash/macro" )



#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" ) 

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 



#   Fetch the current parametrization
rmacroliteRunId( x = par_file ) 
    # [1] 1

#   Modify the parameter
par_file2 <- par_file
rmacroliteRunId( x = par_file2 ) <- 555 

#   Check the result
rid <- rmacroliteRunId( x = par_file2 )
rid 
    # [1] 555 

#   Internal control
if( rid != 555 ){ 
    stop( "Test of rmacroliteRunId() failed" ) } 

#   Clean-up
rm( par_file_path, par_file, par_file2, rid  )

}
