% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteSetModelVar}
\alias{rmacroliteSetModelVar}
\alias{rmacroliteSetModelVar.default}
\title{Set the absolute path of the folder in which MACRO (or MACRO In FOCUS) executable is installed.}
\usage{
rmacroliteSetModelVar(path = NULL, exe = NULL, exeparfile = NULL,
  ...)

\method{rmacroliteSetModelVar}{default}(path = NULL, exe = NULL,
  exeparfile = NULL, ...)
}
\arguments{
\item{path}{Single character string. Absolute path to the folder 
(directory) in which MACRO or MACRO In FOCUS executable 
can be found. When both programs are installed on your 
computer, chose which of the two you want \code{rmacrolite} 
to use. Do not include the name of the executable, 
only its folder. See the introduction above.}

\item{exe}{Single character string. Name, without path, but with the 
extension, of the MACRO or MACRO In FOCUS executable to 
be used by \code{rmacrolite}. The executable must be 
present in the folder \code{path}.  See the introduction 
above.}

\item{exeparfile}{Single character string. Name, without path, but with the 
extension, of the exeparfile executable to 
be used by \code{rmacrolite}. The executable must be 
present in the folder \code{path}. As it is currently 
not shipped with MACRO, the user must install the executable 
beforehand. It can be copy-pasted from a MACRO In FOCUS 
installation. See the introduction above.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Invisibly returns a \code{\link[base]{list}} with 3 named 
 items: \code{path}, \code{exe} and \code{exeparfile}, 
 set to the values found out by the function.
}
\description{
Set the absolute path of the folder in which MACRO (or 
 MACRO In FOCUS) executable is installed. as well as 
 the name of MACRO executable and the name of the exeparfile 
 executable.
 
 Regarding the path to MACRO-folder, the function 
 proceeds as follow: If a value is given (argument 
 code{path}, see below), it is used to set MACRO-path in 
 the package-option \code{macro_path} (see 
 \code{\link[rmacrolite]{rmlPar}}), value that is then used 
 by other function in this package. If no value is given 
 (argument code{path} left to \code{NULL}), the function 
 will first search for a Windows Environment Variable 
 (System Variable) called \code{rmacrolite_macro_path}, 
 and use this value if it exists, to set \code{macro_path} 
 in \code{\link[rmacrolite]{rmlPar}}. See 
 \code{\link[rmacrolite]{rmacrolite-system-variables}}. 
 Finally, if neither the argument 
 code{path} is set, nor a System Variable called 
 \code{rmacrolite_macro_path}, the function will try the 
 factory-default path to MACRO In FOCUS (type 
 \code{getRmlPar("macro_path_default")} to find out).
 
 The principle is the same for MACRO executable (argument 
 \code{exe}, System Variable \code{rmacrolite_macro_exe}, factory 
 default set in \code{getRmlPar("macro_exe_default")}) 
 as well as for the exeparfile executable (argument 
 \code{exeparfile}, System Variable 
 \code{rmacrolite_macro_exeparfile}, factory default set in 
 \code{getRmlPar("macro_exeparfile_default")}).
}
\examples{
# # Not run
# rmacroliteSetModelVar( path = "C:/swash/macro" )
}
\seealso{
\code{\link[rmacrolite]{rmacroliteGetModelVar}} and 
 \code{\link[rmacrolite]{rmacrolite-system-variables}}.
}
