% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteSimPeriod}
\alias{rmacroliteSimPeriod}
\alias{rmacroliteSimPeriod.macroParFile}
\title{Fetch the simulation input period (start / stop time) from imported MACRO parameters}
\usage{
rmacroliteSimPeriod(x, ...)

\method{rmacroliteSimPeriod}{macroParFile}(x, ...)
}
\arguments{
\item{x}{A 'macroParFile' object, such as obtained with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}.}

\item{\dots}{Not used.}
}
\value{
Returns a list with 4 items: \code{sim}, a vector 
 of two \code{\link{POSIXct}} time-dates, start and end 
 time of the simulation; \code{metPeriod}, a vector 
 of two \code{\link{POSIXct}} time-dates, as read from the 
 parameter 
 \code{METPERIOD}. \code{rainBinPeriod} and \code{metBinPeriod} 
 are the same, but read from the rainfall and meteorological 
 data directly.
}
\description{
Fetch the simulation input period (start 
 / stop time) from imported MACRO parameters.
}
\examples{

library( "rmacrolite" )

#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" )  

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 

#   Find out simulation start and end dates
rmacroliteSimPeriod( x = par_file ) 

#   Clean-up
rm( par_file_path, par_file )

}
