% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteSimType}
\alias{rmacroliteSimType}
\alias{rmacroliteSimType.macroParFile}
\alias{rmacroliteSimType<-}
\alias{rmacroliteSimType<-.macroParFile}
\title{Fetch or set the type of simulation in an imported MACRO par-file (parent or metabolite; intermediate or not)}
\usage{
rmacroliteSimType(x, ...)

\method{rmacroliteSimType}{macroParFile}(x, ...)

rmacroliteSimType( x, warn = TRUE, ... ) <- value

\method{rmacroliteSimType}{macroParFile}(x, warn = TRUE, ...) <- value
}
\arguments{
\item{x}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}. This 
file MUST be from a parent substance, and not an 
intermediate simulation output.}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{value}{A \code{\link[base]{list}} with two items. The items 
can be named \code{"type"} and \code{"drivingfile"}. 
The first item (\code{"type"}) should be a single 
integer-value indicating the type of simulation and the 
second item (\code{"drivingfile"}) should be a 
single character-string indicating the name of the 
bin-file to be read-in when simulating a metabolite. 
This file should therefore be an intemediate simulation 
output. When \code{"type"} is set to \code{1} or \code{2} 
(ie a parent substance), the second item can be skipped 
entirely.
The first item (\code{"type"}) indicates the type of 
simulation \code{x} should be changed to: \code{1} is for 
setting the simulation to a parent substance, not intermediate 
simulation output, \code{2} is for changing to a parent 
substance, intermediate simulation output, \code{3} is 
for changing to a metabolite, not intermediate simulation 
output and \code{4} is for changing to a metabolite, 
intermediate simulation output. In practice, as \code{x} 
must be a type \code{1} (parent, not intermediate 
simulation output), setting \code{value} to \code{1} is 
pointless.}
}
\value{
A \code{\link[base]{list}} with two named items.
 The first item (\code{"type"}) is a single integer-value 
 indicating the type of simulation and the second item 
 (\code{"drivingfile"}) is a single character-string 
 indicating the name of the bin-file to be read-in when 
 simulating a metabolite. This file is therefore be an 
 intemediate simulation output. 
 The first item (\code{"type"}) indicates the type of 
 simulation \code{x} presumably contains: \code{1} is for 
 a parent substance, not intermediate simulation output, 
 \code{2} is a parent substance, intermediate simulation 
 output, \code{3} is a metabolite, not intermediate 
 simulation output and \code{4} is a metabolite, intermediate 
 simulation output.
}
\description{
Fetch or set the type of simulation in an imported MACRO 
 par-file (parent or metabolite; intermediate or not). 
 See the description of the argument \code{value} 
 and \code{return} below for a description of the 
 different simulation types. Here, the term the "parent" 
 should be understood as a substance directly applied on 
 the field, not a primary or secondary (etc) metabolite. 
 An intermediate simulation is a simulation that outputs 
 only information on the mass of substance is degraded 
 at each time step and in each numerical layer. 
 Such intermediate simulation is then used as input for 
 simulating the fate of the degradation product(s) of this 
 substance.
}
