% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\docType{methods}
\name{rmacroliteSorption}
\alias{rmacroliteSorption}
\alias{rmacroliteSorption.macroParFile}
\alias{rmacroliteSorption<-}
\alias{rmacroliteSorption<-.macroParFile}
\title{Fetch or set normalised Freundlich adsorption coefficients (Kfoc, Nf) in an imported MACRO par-file}
\usage{
rmacroliteSorption(x, ...)

\method{rmacroliteSorption}{macroParFile}(x, ...)

rmacroliteSorption( x, ... ) <- value

\method{rmacroliteSorption}{macroParFile}(x, ...) <- value
}
\arguments{
\item{x}{A \code{macroParFile}, as imported with 
\code{\link[rmacrolite]{rmacroliteImportParFile-methods}}}

\item{\dots}{Additional parameters passed to specific methods. 
Currently not used.}

\item{value}{Vector of two numeric-value, optionally with the named 
"kfoc" and "nf" (observe the lowercase). 
Value of the Kfoc-coefficient and Nf-exponent to be set 
in the par-file \code{x}, in [m3/kg] and [-], respectively.}
}
\value{
Returns a list with two items, \code{"layer"} 
 (layer-specific parameters) and \code{"site"} 
 (site-specific parameters, i.e. parameters that do 
 not vary with depth or with the crop). \code{"layer"} is 
 columns \code{layer_no}, \code{oc_pc}, \code{kf}, 
 \code{kfoc}, \code{nf}: the layer number, 
 Kf coefficient [m3/kg], the percentage of organic carbon 
 [% dry mass] , the recalculated Kfoc coefficient [m3/kg], 
 the Nf exponent [-], respectively. 
 \code{"site"} is a vector of named numeric-values, with 
 the following item: \code{koc}, the the site-wide 
 non-recalculated kfoc (KOC in MACRO par-files), 
 respectively.
}
\description{
Fetch or set normalised Freundlich adsorption coefficients 
 (Kfoc, Nf) in an imported MACRO par-file. The function 
 fetches both the Kf (non-normalised coefficient; ZKD in 
 MACRO par-files) and the percentage of organic carbon 
 (ORGC in MACRO par-files), recalculates the Kfoc from 
 these values and fetches the Freundlich exponent Nf 
 (also known as 1/n exponent; FREUND in MACRO par-files).
}
\examples{

library( "rmacrolite" ) 

# # Setup MACRO directory (if needed)
# rmacroliteSetModelVar( "C:/swash/macro" )



#   Path to an example par-file
par_file_path <- system.file( "par-files", 
    "chat_winCer_GW-X_900gHa_d182.par", 
    package = "rmacrolite" ) 

#   Import the example par-file
par_file <- rmacroliteImportParFile( 
    file = par_file_path ) 



#   Fetch the current parametrization
rmacroliteSorption( x = par_file ) 
    # $`layer`
      # layer_no oc_pc   kf kfoc   nf
    # 1        1  1.39 1.39  100 0.95
    # 2        2  0.93 0.93  100 0.95
    # 3        3  0.70 0.70  100 0.95
    # 4        4  0.30 0.30  100 0.95
    # 5        5  0.30 0.30  100 0.95
    # 6        6  0.27 0.27  100 0.95

    # $site
    # koc 
    # 100

#   Modify the parameter
par_file2 <- par_file
rmacroliteSorption( x = par_file2 ) <- c( "kfoc" = 200, 
    "nf" = 0.90 )

#   Check the result
ads <- rmacroliteSorption( x = par_file2 )
ads 
    # $`layer`
      # layer_no oc_pc   kf kfoc  nf
    # 1        1  1.39 2.78  200 0.9
    # 2        2  0.93 1.86  200 0.9
    # 3        3  0.70 1.40  200 0.9
    # 4        4  0.30 0.60  200 0.9
    # 5        5  0.30 0.60  200 0.9
    # 6        6  0.27 0.54  200 0.9

    # $site
    # koc 
    # 200

#   Internal control
if( !all( ads[[ "layer" ]][, "kfoc" ] == 200 ) ){ 
    stop( "Test of rmacroliteSorption() failed (kfoc, 1)" ) } 

if( ads[[ "site" ]][ "koc" ] != 200  ){ 
    stop( "Test of rmacroliteSorption() failed (kfoc, 2)" ) } 

if( !all( ads[[ "layer" ]][, "nf" ] == 0.9 ) ){ 
    stop( "Test of rmacroliteSorption() failed (nf)" ) } 

#   Clean-up
rm( par_file_path, par_file, par_file2, ads  )

}
