% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrolite.r
\name{.rml_textToPOSIXct}
\alias{.rml_textToPOSIXct}
\title{Convert a text representation of a date to as POSIXct-class object}
\usage{
.rml_textToPOSIXct(x, format = getRmlPar("timeFormat"),
  tz = getRmlPar("tz"), ...)
}
\arguments{
\item{x}{See \code{\link[base:as.POSIXlt]{as.POSIXct}}. Can be a vector of character 
strings}

\item{format}{See \code{\link[base:as.POSIXlt]{as.POSIXct}}.}

\item{\dots}{See \code{\link[base:as.POSIXlt]{as.POSIXct}}.}
}
\value{
Returns the result of 
 \code{\link[base:as.POSIXlt]{as.POSIXct}}\code{(x=x,format=format,tz=tz,...)} 
 and gives warning if thinks the dates are ill-formatted.
}
\description{
Convert a text representation of a date to as POSIXct-class object.
 Performs additional checks as compared to 
 \code{\link[base:as.POSIXlt]{as.POSIXct}} (on which the function is build)
}
\keyword{internal}
